pyproject.toml
[tool.poetry]
name = "siameseattn"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
#torch = { url = "https://download.pytorch.org/whl/cu102/torch-1.8.1%2Bcu102-cp39-cp39-linux_x86_64.whl" }
spacy = "^3.0.5"
numpy = "^1.20.2"
matplotlib = "^3.4.1"
optuna = "^2.7.0"
pynvml = "^8.0.4"
sentencepiece = "^0.1.95"
pytorch-nlp = "^0.5.0"
fvcore = "^0.1.5"
tqdm = "^4.61.0"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
